+++ test/unit-test/scripts/npm/prepublish/lib/build-spec.js
@@ -7,7 +7,7 @@ let execStub;
 
 describe( 'build', () => {
   beforeEach( () => {
-    execStub = jest.spyOn( childProcess, 'exec' );
+    execStub = jest.spyOn( childProcess, 'exec' ).mockImplementation();
   } );
 
   afterEach( () => {