+++ CHANGELOG.md
@@ -9,6 +9,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 ### Changed
 - **cf-core:** [MINOR] Move less breakpoint variables to their own file.
 - **cf-core:** [PATCH] Fixed missing comma in cf-utilities media queries.
+- **capital-framework:** [PATCH] Mock child process methods in script unit tests to prevent erroneous package publishing.
 
 ### Removed
 -